tasklist post

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();
+        if(\Auth::check()) {
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // メッセージ一覧を取得
+            $tasks = $user->taskusers()->orderBy('created_at', 'desc')->paginate(10);
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            // メッセージ一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
+        
     }
 
     /**
@@ -51,14 +56,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:255',
         ]);
-        // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        
+        $request->user()->taskusers()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('tasks');
     }
 
     /**
@@ -109,15 +114,19 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::find($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            
+        }
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('tasks');
     }
 
     /**
@@ -128,12 +137,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::find($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('tasks');
     }
 }
